pyproject.toml
[project]
name = "fastapi-contacts-api"
version = "0.1.0"
description = ""
authors = [
    {name = "TStakhniuk",email = "tpstakhniuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "faker (>=37.6.0,<38.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-faker (>=2.0.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

